Create Ripple.mediawiki
I am aware that this short description is way too long (About 3000 characters). Nevertheless, I ask this pull request to go through this text and to give suggestions for a possible text shortening. My goal was to bring a clear description into the somewhat confusing construct of Ripple. Another possibility would be to outsource parts of the description to "additional information", what do you think?