Create EtherInc.mediawiki
Thank you for reviewing the company profile of EtherInc. 
I was not sure what to write in the "Tradeability" section. The eInc blockchain is a fork from the Ethereum Main Network. So the native currency on this blockchain could be considered a coin, like Ether. Do we want to differentiate between coin and token and if so, how?