Update modum.io.mediawiki
Thank you for your review @joschicif ! Concerning the revenue model I changed the tag to "physical sales". In this way we do not have to define if the company is manufacturing the physical products themselves, which would be the case if we used "production" or "retail".  I chose "physical" because I had the feeling this word would describe it best. Other options would be material or tangible.
What do you think of my solution? @Chainomics and @Fabiansk please comment on this topic as well.